requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
jinja2==3.1.4
python-dotenv==1.0.1
httpx==0.27.0
anthropic==0.42.0
atlassian-python-api==3.41.0
pydantic==2.9.0
python-multipart==0.0.9
markdown==3.7
pyyaml==6.0.2
schedule==1.2.2
rich==13.9.0
click==8.1.7
python-docx==1.1.2
weasyprint==62.3
posthog==3.7.0
google-analytics-data==0.18.0
klaviyo-api==9.0.0
